feat: test for products users and cart

+++ src/dao/mongo/controllers/product.controller.mongo.js
@@ -37,6 +37,9 @@ export class ProductController {
 
   async getById(id) {
     try {
+      if(id.length !== 24) {
+        throw new Error("Bad request: invalid id");
+      }
       const result = await this.model.findById(id);
       if (!result) {
         throw new Error(`Product not found: ${id}`);
@@ -49,6 +52,9 @@ export class ProductController {
 
   async update(id, product) {
     try {
+      if(id.length !== 24) {
+        throw new Error("Bad request: invalid id");
+      }
       const result = await this.model.findByIdAndUpdate(id, product, { new: true });
       if (!result) {
         throw new Error(`Product not found: ${id}`);
@@ -61,6 +67,9 @@ export class ProductController {
 
   async delete(id) {
     try {
+      if(id.length !== 24) {
+        throw new Error("Bad request: invalid id");
+      }
       const result = await this.model.findByIdAndDelete(id);
       if (!result) {
         throw new Error(`Product not found: ${id}`);
